feat: separate dataset generation from model building (#64)
* feat: separate dataset creation from model build

* fix: remove unsupported operations from integration test

* fix: bugs in the datasetgenerator class

* fix: update unit tests for data gen changes

* feat: update readme to reflect data gen changes

* fix: weird import statement in readme

* chore: run black pre-commit

* fix: remove unused import from resolver.py

* chore: bump version to 0.8.0

* feat: change model contract to take list of datasets

A dict of datasets seemed clunky.

* fix: set encoding to utf-8 when writing code file

* chore: split integration tests

+++ smolmodels/fileio.py
@@ -137,13 +137,15 @@ def load_model(path: str) -> Model:
         # Restore artifacts
         if temp_dir.exists():
             for artifact_path in temp_dir.iterdir():
-                shutil.copy(artifact_path, model.files_path / artifact_path.name)
-                model.artifacts.append(str(model.files_path / artifact_path.name))
+                if artifact_path.is_file():
+                    shutil.copy(artifact_path, model.files_path / artifact_path.name)
+                    model.artifacts.append(str(model.files_path / artifact_path.name))
+                elif artifact_path.is_dir():
+                    shutil.copytree(artifact_path, model.files_path / artifact_path.name)
+                    model.artifacts.append(str(model.files_path / artifact_path.name))
 
         with open(trainer_path, "r") as f:
-            model.trainer = types.ModuleType("trainer")
             model.trainer_source = f.read()
-            exec(model.trainer_source, model.trainer.__dict__)
 
         with open(predictor_path, "r") as f:
             model.predictor = types.ModuleType("predictor")
